Перенос отрисовки точек маршрута в Презентер
